EUCA-4597, EUCA-4915: haproxy interface and the additional service client for loadbalancer VMs.
The servo python service, running inside the loadbalancer VMs periodically call desribe-load-balancers,
and reconfigure the haproxy process based on the listener specification it receives from the loadbalancer service.
Note that the request to describe-loadbalancer contains a marker which has the instance id of the requesting LB VM,
and the service implementation should return only the listener specification that the LB VM is entitiled to forward.